+++ src/lib/tls/asio/info.txt
@@ -13,8 +13,6 @@ asio_async_base.h
 asio_async_handshake_op.h
 asio_async_read_op.h
 asio_async_write_op.h
-asio_stream_base.h
-asio_stream_core.h
 asio_includes.h
 </header:internal>
